+++ template/runtime/src/lib.rs
@@ -295,6 +295,7 @@ impl pallet_evm::Config for Runtime {
 		pallet_evm_precompile_simple::Sha256,
 		pallet_evm_precompile_simple::Ripemd160,
 		pallet_evm_precompile_simple::Identity,
+		pallet_evm_precompile_modexp::Modexp,
 		pallet_evm_precompile_simple::ECRecoverPublicKey,
 		pallet_evm_precompile_sha3fips::Sha3FIPS256,
 		pallet_evm_precompile_sha3fips::Sha3FIPS512,
@@ -557,21 +558,21 @@ impl_runtime_apis! {
 
 		fn current_transaction_statuses() -> Option<Vec<TransactionStatus>> {
 			match Ethereum::current_transaction_statuses() {
-				Some(elt) => Some(elt.into()),
+				Some(elt) => Some(elt.inner.into()),
 				None => None,
 			}
 		}
 
 		fn current_block() -> Option<pallet_ethereum::Block> {
 			match Ethereum::current_block() {
-				Some(elt) => Some(elt.into()),
+				Some(elt) => Some(elt.inner.into()),
 				None => None,
 			}
 		}
 
 		fn current_receipts() -> Option<Vec<pallet_ethereum::Receipt>> {
 			match Ethereum::current_receipts() {
-				Some(elt) => Some(elt.into()),
+				Some(elt) => Some(elt.inner.into()),
 				None => None,
 			}
 		}