+++ client/rpc-core/CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog for `fc-rpc-core`
 
 ## Unreleased
+- Add `FilteredParams::in_bloom()` function to check the possible existance of Filter addresses or topics in a block.